app/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
config==0.5.1
discord.py==2.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.4
mutagen==1.47.0
pycparser==2.21
PyNaCl==1.5.0
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp==2023.12.30
